+++ src/main/java/com/vamos/services/EnderecoService.java
@@ -10,7 +10,7 @@ public class EnderecoService {
 	
 	private EnderecoRepository enderecoRepository;
 	
-	public Endereco find(long id) {
+	public Endereco find(Integer id) {
 		Optional<Endereco> obj = enderecoRepository.findById(id);
 		return obj.orElseThrow(() -> new ObjectNotFoundException("Endereco não encontrado! Id: " + id + ", Tipo: " + Endereco.class.getName()));
 	}
